+++ Assets/Scripts/UI/UILobbiesMenu.cs
@@ -47,6 +47,7 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
         public Toggle AllowInvitesVal;
         public Toggle PresenceEnabledVal;
         public Toggle RTCVoiceRoomEnabledVal;
+        public Toggle AntiCheatEnabledVal;
 
         // Create/Modify/Leave UI
         public Button CreateLobbyButton;
@@ -88,6 +89,7 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
 
         private EOSLobbyManager LobbyManager;
         private EOSFriendsManager FriendsManager;
+        private EOSEACLobbyManager AntiCheatLobbyManager;
 
         public void Awake()
         {
@@ -105,6 +107,39 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
 
             LobbyManager = EOSManager.Instance.GetOrCreateManager<EOSLobbyManager>();
             FriendsManager = EOSManager.Instance.GetOrCreateManager<EOSFriendsManager>();
+            AntiCheatLobbyManager = EOSManager.Instance.GetOrCreateManager<EOSEACLobbyManager>();
+            if (AntiCheatEnabledVal != null)
+            {
+                AntiCheatEnabledVal.isOn = false;
+                AntiCheatEnabledVal.interactable = AntiCheatLobbyManager.IsAntiCheatAvailable();
+            }
+
+            LobbyManager.SubscribeToMemberUpdates(OnMemberUpdate);
+        }
+
+        private void OnDestroy()
+        {
+            LobbyManager?.UnsubscribeFromMemberUpdates(OnMemberUpdate);
+        }
+
+        private void OnMemberUpdate(string LobbyId, ProductUserId MemberId)
+        {
+            Lobby currentLobby = LobbyManager.GetCurrentLobby();
+            if (currentLobby.Id != LobbyId)
+            {
+                return;
+            }
+
+            UIMemberEntry uiEntry = UIMemberEntries.Find((UIMemberEntry entry) => { return entry.ProductUserId == MemberId; });
+            if (uiEntry != null)
+            {
+                LobbyMember updatedMember = currentLobby.Members.Find((LobbyMember member) => { return member.ProductId == MemberId; });
+                if (updatedMember != null)
+                {
+                    uiEntry.UpdateMemberData(updatedMember);
+                    uiEntry.UpdateUI();
+                }
+            }
         }
 
         private void Update()
@@ -191,25 +226,8 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
                         UIMemberEntry uiEntry = memberUIObj.GetComponent<UIMemberEntry>();
                         if (uiEntry != null)
                         {
-                            string displayName = member.DisplayName;
-                            if (!string.IsNullOrEmpty(displayName))
-                            {
-                                uiEntry.MemberName = displayName;
-                            }
-                            else
-                            {
-                                Result result = member.ProductId.ToString(out Utf8String outBuff);
-                                if (result == Result.Success)
-                                {
-                                    uiEntry.MemberName = outBuff;
-                                }
-                                else
-                                {
-                                    uiEntry.MemberName = "Error: " + result;
-                                }
-                            }
+                            uiEntry.UpdateMemberData(member);
 
-                            uiEntry.ProductUserId = member.ProductId;
                             //uiEntry.IsOwner = currentLobby.LobbyOwner == member.ProductId;
 
                             uiEntry.IsTalkingText.text = "---";
@@ -293,12 +311,20 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
             lobbyProperties.MaxNumLobbyMembers = (uint)Int32.Parse(MaxPlayersVal.options[MaxPlayersVal.value].text);
 
             // Level Attributes
-            LobbyAttribute attribute = new LobbyAttribute();
-            attribute.Key = "LEVEL";
-            attribute.AsString = LevelVal.options[LevelVal.value].text;
-            attribute.ValueType = AttributeType.String;
-            attribute.Visibility = LobbyAttributeVisibility.Public;  // Needs to be public for Search
-            lobbyProperties.Attributes.Add(attribute);
+            LobbyAttribute levelAttribute = new LobbyAttribute();
+            levelAttribute.Key = "LEVEL";
+            levelAttribute.AsString = LevelVal.options[LevelVal.value].text;
+            levelAttribute.ValueType = AttributeType.String;
+            levelAttribute.Visibility = LobbyAttributeVisibility.Public;  // Needs to be public for Search
+            lobbyProperties.Attributes.Add(levelAttribute);
+
+            // Anti-Cheat Attributes
+            LobbyAttribute antiCheatAttribute = new LobbyAttribute();
+            antiCheatAttribute.Key = "ANTICHEAT";
+            antiCheatAttribute.AsBool = AntiCheatEnabledVal.isOn;
+            antiCheatAttribute.ValueType = AttributeType.Boolean;
+            antiCheatAttribute.Visibility = LobbyAttributeVisibility.Public;
+            lobbyProperties.Attributes.Add(antiCheatAttribute);
 
             // Permission
             string permissionStr = PermissionVal.options[PermissionVal.value].text;